feat(flags): support root, force, linked, and owner flags
Support a global root flag to change the root directory from the system
root. This prepares for writing e2e tests where we want to specify where
files are written. Support a link force flag which allows setting a link
even when the link already exists. This is useful for downgrading
permissions of a link, like changing the owner from root to something
else. Update naming of the list link flag to linked. Support a list
owner flag which lists the user owner with a link.

General refactoring and fixing related to flag processing. Given the way
we process flags right now, global flags need to be skipped when
processing subcommand flags. The converse is also true, where subcommand
flags need to be skipped when processing global flags. We also expose a
few new global variables which make passing options easier.